+++ frontend/src/components/pages/UserCodeReviewPerformance/theme.js
@@ -1,6 +1,6 @@
-import { createTheme } from "@material-ui/core/styles";
+import { createMuiTheme } from "@material-ui/core/styles";
 
-const theme = createTheme({
+const theme = createMuiTheme({
   overrides: {
     MuiTooltip: {
       tooltip: {
