--- frontend/src/components/widgets/theme.js
@@ -1,13 +1,0 @@
-import { createTheme } from '@material-ui/core/styles';
-
-const theme = createTheme({
-    overrides: {
-        MuiTooltip: {
-            tooltip: {
-                fontSize: "1em",
-            },
-        },
-    },
-});
-
-export default theme;
